chore(core): fix xframebuffer conditionals
[no changelog]
