+++ core/embed/rust/src/ui/shape/mod.rs
@@ -10,6 +10,8 @@ mod corner_highlight;
 mod display;
 #[cfg(feature = "ui_jpeg_decoder")]
 mod jpeg;
+#[cfg(not(feature = "xframebuffer"))]
+mod progressive_render;
 mod qrcode;
 mod rawimage;
 mod render;
@@ -31,9 +33,12 @@ pub use corner_highlight::CornerHighlight;
 pub use display::{render_on_canvas, render_on_display, unlock_bumps_on_failure, ConcreteRenderer};
 #[cfg(feature = "ui_jpeg_decoder")]
 pub use jpeg::JpegImage;
+#[cfg(not(feature = "xframebuffer"))]
+pub use progressive_render::ProgressiveRenderer;
 pub use qrcode::QrImage;
 pub use rawimage::RawImage;
-pub use render::{DirectRenderer, ProgressiveRenderer, Renderer, ScopedRenderer};
+pub use render::{DirectRenderer, Renderer, ScopedRenderer};
+
 pub use text::Text;
 pub use toif::ToifImage;
 #[cfg(feature = "ui_image_buffer")]
